int to float in subject mark

+++ TestFactory.Business/Models/SubjectMark.cs
@@ -9,6 +9,6 @@ namespace TestFactory.Business.Models
         public virtual string StudentId { get; set; }
 
         [Required]
-        public virtual int? Value { get; set; }
+        public virtual float? Value { get; set; }
     }
 }
